+++ tests/cpp/core/RestServerTests.cpp
@@ -42,12 +42,13 @@
 
 #include "rest/RestServer.h"
 
-#include <zeroeq/uri.h>
-
 #include <QNetworkAccessManager>
 #include <QNetworkReply>
 
 #include "MinimalGlobalQtApp.h"
+
+using namespace zeroeq;
+
 BOOST_GLOBAL_FIXTURE(MinimalGlobalQtApp);
 
 BOOST_AUTO_TEST_CASE(testDefaultPort)
@@ -62,7 +63,8 @@ BOOST_AUTO_TEST_CASE(testUnavailablePort)
     BOOST_CHECK_THROW(RestServer server{80}, std::runtime_error);
 }
 
-QString performHttpRequest(const QUrl& url)
+std::pair<std::string, int> sendHttpRequest(const QUrl& url,
+                                            const http::Method method)
 {
     // create custom temporary event loop on stack
     QEventLoop eventLoop;
@@ -71,11 +73,39 @@ QString performHttpRequest(const QUrl& url)
     QObject::connect(&manager, &QNetworkAccessManager::finished, &eventLoop,
                      &QEventLoop::quit);
 
-    std::unique_ptr<QNetworkReply> reply(manager.get(QNetworkRequest{url}));
+    std::unique_ptr<QNetworkReply> reply;
+    switch (method)
+    {
+    case http::Method::GET:
+        reply.reset(manager.get(QNetworkRequest{url}));
+        break;
+    case http::Method::POST:
+        reply.reset(manager.post(QNetworkRequest{url}, ""));
+        break;
+    case http::Method::PUT:
+        reply.reset(manager.put(QNetworkRequest{url}, ""));
+        break;
+    case http::Method::OPTIONS:
+        reply.reset(manager.sendCustomRequest(QNetworkRequest{url}, "OPTIONS"));
+        break;
+    case http::Method::DELETE:
+        reply.reset(manager.deleteResource(QNetworkRequest{url}));
+        break;
+    case http::Method::PATCH:
+        reply.reset(manager.sendCustomRequest(QNetworkRequest{url}, "PATCH"));
+        break;
+    case http::Method::ALL:
+        break;
+    }
+
     eventLoop.exec(); // blocks stack until "finished()" has been called
 
-    return (reply->error() == QNetworkReply::NoError) ? reply->readAll()
-                                                      : reply->errorString();
+    const auto response = (reply->error() == QNetworkReply::NoError)
+                              ? reply->readAll()
+                              : reply->errorString();
+
+    return std::pair<std::string, int>(response.toStdString(),
+                                       int(reply->error()));
 }
 
 struct TestObject
@@ -96,7 +126,97 @@ BOOST_AUTO_TEST_CASE(testServerReturnsSimpleContent)
     server.handleGET("test", test);
 
     const auto url = QString("http://localhost:%1/test").arg(server.getPort());
-    const auto response = performHttpRequest(url);
+    const auto response = sendHttpRequest(url, http::Method::GET);
 
-    BOOST_CHECK_EQUAL(response.toStdString(), "Hello World!");
+    BOOST_CHECK_EQUAL(response.first, "Hello World!");
+    BOOST_CHECK_EQUAL(response.second, int(QNetworkReply::NoError));
+}
+
+class MockRestServer : public RestServer
+{
+public:
+    bool skipWhitelistCheck = true;
+
+    // Exposing private respondTo
+    std::future<http::Response> mockRespondTo(
+        zeroeq::http::Request& request) const
+    {
+        return RestServer::respondTo(request);
+    }
+
+private:
+    // Bypass hardcoded localhost exception for block()
+    bool _isWhitelisted(const std::string&) const final
+    {
+        return skipWhitelistCheck;
+    }
+};
+
+BOOST_AUTO_TEST_CASE(block_all_methods)
+{
+    MockRestServer server;
+
+    for (int method = 0; method < int(zeroeq::http::Method::ALL); ++method)
+    {
+        server.handle(zeroeq::http::Method(method), "test",
+                      [](const http::Request&) {
+                          return http::make_ready_response(http::Code::OK);
+                      });
+    };
+
+    const auto url = QString("http://localhost:%1/test").arg(server.getPort());
+
+    for (int method = 0; method < int(zeroeq::http::Method::ALL); ++method)
+    {
+        const auto response =
+            sendHttpRequest(url, zeroeq::http::Method(method));
+        BOOST_CHECK_EQUAL(response.second, int(QNetworkReply::NoError));
+    }
+
+    server.skipWhitelistCheck = false;
+
+    for (int method = 0; method < int(zeroeq::http::Method::ALL); ++method)
+    {
+        server.block(zeroeq::http::Method(method));
+        const auto response =
+            sendHttpRequest(url, zeroeq::http::Method(method));
+        // localhost is no longer whitelisted because of bypassWhitelist flag.
+        BOOST_CHECK_EQUAL(
+            response.second,
+            int(QNetworkReply::ContentOperationNotPermittedError));
+    }
+}
+
+BOOST_AUTO_TEST_CASE(test_whitelist)
+{
+    MockRestServer server;
+    http::Request localhostRequest;
+    localhostRequest.method = http::Method::PUT;
+    localhostRequest.path = "/test";
+    localhostRequest.source = "127.0.0.1";
+    localhostRequest.body = "das";
+
+    http::Request foreingRequest;
+    foreingRequest.method = http::Method::PUT;
+    foreingRequest.path = "/test";
+    foreingRequest.source = "172.16.0.1";
+    foreingRequest.body = "das";
+
+    server.block(http::Method::PUT);
+
+    server.handle(zeroeq::http::Method::PUT, "test", [](const http::Request&) {
+        return http::make_ready_response(http::Code::OK);
+    });
+
+    server.skipWhitelistCheck = true;
+
+    const auto url = QString("http://localhost:%1/test").arg(server.getPort());
+
+    auto response = server.mockRespondTo(localhostRequest);
+    BOOST_CHECK_EQUAL(response.get().code, http::Code::OK);
+
+    server.skipWhitelistCheck = false;
+
+    response = server.mockRespondTo(foreingRequest);
+    BOOST_CHECK_EQUAL(response.get().code, http::Code::FORBIDDEN);
 }